+++ server/Dockerfile.dev
@@ -1,0 +1,22 @@
+FROM node:20 as builder
+
+WORKDIR /myapp
+COPY package.json .
+RUN npm install
+
+COPY . .
+RUN npm run build
+
+# Etapa 2:
+
+FROM node:20-alpine
+
+WORKDIR /myapp
+
+COPY --from=builder /myapp/dist ./dist
+COPY --from=builder /myapp/node_modules ./node_modules
+COPY --from=builder /myapp/package*.json ./
+
+EXPOSE 4000
+
+CMD ["npm", "start"]
