+++ server/src/swagger.config.ts
@@ -1,0 +1,14 @@
+import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+import { INestApplication } from '@nestjs/common';
+
+export default function setupSwagger(app: INestApplication): void {
+  const config = new DocumentBuilder()
+    .setTitle('Blogs example')
+    .addBearerAuth()
+    .setDescription('The blogs API description')
+    .setVersion('1.0')
+    .addTag('blogs')
+    .build();
+  const document = SwaggerModule.createDocument(app, config);
+  SwaggerModule.setup('docs', app, document);
+}
